move GetConsent() to common/user_utils, linter fixes and minor CR updates
